Refactor : post컴포넌트 구조 변경
